move get_mode_info to default_plugin
