+++ src/common/input/handler.rs
@@ -1,8 +1,7 @@
 //! Main input logic.
 
 use super::keybinds::Keybinds;
-use super::{actions::Action, keybinds::ModeKeybinds};
-use crate::common::input::{actions::Direction, config::Config};
+use crate::common::input::{config::Config};
 use crate::common::{AppInstruction, SenderWithContext, OPENCALLS};
 use crate::errors::ContextType;
 use crate::os_input_output::OsApi;
@@ -12,7 +11,7 @@ use crate::wasm_vm::PluginInstruction;
 use crate::CommandIsExecuting;
 
 use termion::input::{TermRead, TermReadEventsAndRaw};
-use zellij_tile::data::{Event, InputMode, Key, ModeInfo};
+use zellij_tile::{actions::{Action, Direction}, data::{Event, InputMode, Key, ModeInfo}};
 
 /// Handles the dispatching of [`Action`]s according to the current
 /// [`InputMode`], and keep tracks of the current [`InputMode`].
@@ -128,17 +127,20 @@ impl InputHandler {
             }
             Action::SwitchToMode(mode) => {
                 self.mode = mode;
+                let keybinds = self.config.keybinds.0.get(&mode).cloned().unwrap_or(Keybinds::get_defaults_for_mode(&mode)).0;
                 self.send_plugin_instructions
                     .send(PluginInstruction::Update(
                         None,
-                        Event::ModeUpdate(get_mode_info(mode, &self.config.keybinds)),
+                        Event::ModeUpdate(ModeInfo {
+                            mode: mode.clone(), 
+                            keybinds: keybinds.clone()}),
                     ))
                     .unwrap();
                 self.send_screen_instructions
-                    .send(ScreenInstruction::ChangeMode(get_mode_info(
+                    .send(ScreenInstruction::ChangeMode(ModeInfo{
                         mode,
-                        &self.config.keybinds,
-                    )))
+                        keybinds,
+                    }))
                     .unwrap();
                 self.send_screen_instructions
                     .send(ScreenInstruction::Render)
@@ -146,10 +148,10 @@ impl InputHandler {
             }
             Action::Resize(direction) => {
                 let screen_instr = match direction {
-                    super::actions::Direction::Left => ScreenInstruction::ResizeLeft,
-                    super::actions::Direction::Right => ScreenInstruction::ResizeRight,
-                    super::actions::Direction::Up => ScreenInstruction::ResizeUp,
-                    super::actions::Direction::Down => ScreenInstruction::ResizeDown,
+                    Direction::Left => ScreenInstruction::ResizeLeft,
+                    Direction::Right => ScreenInstruction::ResizeRight,
+                    Direction::Up => ScreenInstruction::ResizeUp,
+                    Direction::Down => ScreenInstruction::ResizeDown,
                 };
                 self.send_screen_instructions.send(screen_instr).unwrap();
             }
@@ -170,10 +172,10 @@ impl InputHandler {
             }
             Action::MoveFocus(direction) => {
                 let screen_instr = match direction {
-                    super::actions::Direction::Left => ScreenInstruction::MoveFocusLeft,
-                    super::actions::Direction::Right => ScreenInstruction::MoveFocusRight,
-                    super::actions::Direction::Up => ScreenInstruction::MoveFocusUp,
-                    super::actions::Direction::Down => ScreenInstruction::MoveFocusDown,
+                    Direction::Left => ScreenInstruction::MoveFocusLeft,
+                    Direction::Right => ScreenInstruction::MoveFocusRight,
+                    Direction::Up => ScreenInstruction::MoveFocusUp,
+                    Direction::Down => ScreenInstruction::MoveFocusDown,
                 };
                 self.send_screen_instructions.send(screen_instr).unwrap();
             }
@@ -204,16 +206,16 @@ impl InputHandler {
             }
             Action::NewPane(direction) => {
                 let pty_instr = match direction {
-                    Some(super::actions::Direction::Left) => {
+                    Some(Direction::Left) => {
                         PtyInstruction::SpawnTerminalVertically(None)
                     }
-                    Some(super::actions::Direction::Right) => {
+                    Some(Direction::Right) => {
                         PtyInstruction::SpawnTerminalVertically(None)
                     }
-                    Some(super::actions::Direction::Up) => {
+                    Some(Direction::Up) => {
                         PtyInstruction::SpawnTerminalHorizontally(None)
                     }
-                    Some(super::actions::Direction::Down) => {
+                    Some(Direction::Down) => {
                         PtyInstruction::SpawnTerminalHorizontally(None)
                     }
                     // No direction specified - try to put it in the biggest available spot
@@ -282,195 +284,6 @@ impl InputHandler {
             .send(AppInstruction::Exit)
             .unwrap();
     }
-}
-
-// const fn now does not support PartialEq/Eq, we have to implement our own compare fn
-const fn compare_key(l: &Key, r: &Key) -> bool {
-    match (l, r) {
-        (Key::Backspace, Key::Backspace) |
-        (Key::Left, Key::Left) |
-        (Key::Right, Key::Right) |
-        (Key::Up, Key::Up) |
-        (Key::Down, Key::Down) |
-        (Key::Home, Key::Home) |
-        (Key::End, Key::End) |
-        (Key::PageUp, Key::PageUp) |
-        (Key::PageDown, Key::PageDown) |
-        (Key::Delete, Key::Delete) |
-        (Key::Insert, Key::Insert) |
-        (Key::Esc, Key::Esc) |
-        (Key::BackTab, Key::BackTab) => true,
-        _ => false,
-    }
-}
-
-const fn get_key_order(key: &Key) -> Option<i32> {
-    const V : &[(Key, i32)]= &[
-        (Key::Left, 0),
-        (Key::Right, 0),
-        (Key::Up, 1),
-        (Key::Down, 1),
-        (Key::PageUp, 2),
-        (Key::PageDown, 2),
-    ];
-    let mut i = 0;
-    while  i < V.len() {
-        let (k, o) = V[i];
-        if compare_key(&k, key){
-            return Some(o);
-        }
-        i += 1;
-    }
-    None
-}
-
-/// Get a prior key from keybinds
-/// many keys may be mapped to one action, e.g. kj/↑↓
-/// but we do not want to show all of them in help info,
-/// so just pickup one primary key.
-fn get_major_key_by_action(keybinds: &ModeKeybinds, action: &[Action]) -> Key {
-    let mut key = Key::Null;
-    for (k, actions) in &keybinds.0 {
-        if actions == action {
-            if key == Key::Null {
-                // old key is null
-                key = *k;
-            } else if let Some(new_order) = get_key_order(k) {
-                if let Some(old_order) = get_key_order(&key) {
-                    if new_order < old_order {
-                        // old key has lower order (larger number) than new one
-                        key = *k;
-                    }
-                } else {
-                    // old key does not have order, new key have order
-                    // then use new keybind
-                    key = *k;
-                }
-            }
-        }
-    }
-    key
-}
-
-fn get_key_map_string(key_config: &ModeKeybinds, actions: &[&[Action]]) -> String {
-    let map = actions
-        .iter()
-        .map(|&actions| get_major_key_by_action(&key_config, actions))
-        .map(|key| key.to_string())
-        .collect::<Vec<_>>();
-    let should_split = map.iter().any(|s| s.chars().count() > 1);
-    map.into_iter().fold(String::new(), |s0, s| {
-        if !s0.is_empty() && should_split {
-            format!("{}/{}", s0, s)
-        } else {
-            format!("{}{}", s0, s)
-        }
-    })
-}
-
-/// Creates a [`Help`] struct indicating the current [`InputMode`] and its keybinds
-/// (as pairs of [`String`]s).
-// TODO this should probably be automatically generated in some way
-pub fn get_mode_info(mode: InputMode, key_config: &Keybinds) -> ModeInfo {
-    let key_config = key_config
-        .0
-        .get(&mode)
-        .cloned()
-        .unwrap_or_else(|| Keybinds::get_defaults_for_mode(&mode));
-    let mut keybinds: Vec<(String, String)> = vec![];
-    match mode {
-        InputMode::Normal | InputMode::Locked => {}
-        InputMode::Resize => {
-            let key_map = get_key_map_string(
-                &key_config,
-                &[
-                    &[Action::Resize(Direction::Left)],
-                    &[Action::Resize(Direction::Down)],
-                    &[Action::Resize(Direction::Up)],
-                    &[Action::Resize(Direction::Right)],
-                ],
-            );
-            keybinds.push((key_map, "Resize".to_string()));
-        }
-        InputMode::Pane => {
-            let key_map = get_key_map_string(
-                &key_config,
-                &[
-                    &[Action::MoveFocus(Direction::Left)],
-                    &[Action::MoveFocus(Direction::Down)],
-                    &[Action::MoveFocus(Direction::Up)],
-                    &[Action::MoveFocus(Direction::Right)],
-                ],
-            );
-            keybinds.push((key_map, "Move focus".to_string()));
-            keybinds.push((
-                get_major_key_by_action(&key_config, &[Action::SwitchFocus]).to_string(),
-                "Next".to_string(),
-            ));
-            keybinds.push((
-                get_major_key_by_action(&key_config, &[Action::NewPane(None)]).to_string(),
-                "New".to_string(),
-            ));
-            keybinds.push((
-                get_major_key_by_action(&key_config, &[Action::NewPane(Some(Direction::Down))])
-                    .to_string(),
-                "Down split".to_string(),
-            ));
-            keybinds.push((
-                get_major_key_by_action(&key_config, &[Action::NewPane(Some(Direction::Right))])
-                    .to_string(),
-                "Right split".to_string(),
-            ));
-            keybinds.push((
-                get_major_key_by_action(&key_config, &[Action::CloseFocus]).to_string(),
-                "Close".to_string(),
-            ));
-            keybinds.push((
-                get_major_key_by_action(&key_config, &[Action::ToggleFocusFullscreen]).to_string(),
-                "Fullscreen".to_string(),
-            ));
-        }
-        InputMode::Tab => {
-            let key_map = get_key_map_string(
-                &key_config,
-                &[&[Action::GoToPreviousTab], &[Action::GoToNextTab]],
-            );
-            keybinds.push((key_map, "Move focus".to_string()));
-            keybinds.push((
-                get_major_key_by_action(&key_config, &[Action::NewTab]).to_string(),
-                "New".to_string(),
-            ));
-            keybinds.push((
-                get_major_key_by_action(&key_config, &[Action::CloseTab]).to_string(),
-                "Close".to_string(),
-            ));
-            keybinds.push((
-                get_major_key_by_action(
-                    &key_config,
-                    &[
-                        Action::SwitchToMode(InputMode::RenameTab),
-                        Action::TabNameInput(vec![0]),
-                    ],
-                )
-                .to_string(),
-                "Rename".to_string(),
-            ));
-        }
-        InputMode::Scroll => {
-            let key_map =
-                get_key_map_string(&key_config, &[&[Action::ScrollUp], &[Action::ScrollDown]]);
-            keybinds.push((key_map, "Scroll".to_string()));
-            let key_map = get_key_map_string(
-                &key_config,
-                &[&[Action::PageScrollUp], &[Action::PageScrollDown]],
-            );
-            keybinds.push((key_map, "Scroll Page".to_string()));
-        }
-        InputMode::RenameTab => {
-            keybinds.push(("Enter".to_string(), "when done".to_string()));
-        }
-    }
-    ModeInfo { mode, keybinds }
 }
 
 /// Entry point to the module. Instantiates an [`InputHandler`] and starts
