+++ src/common/input/keybinds.rs
@@ -1,12 +1,9 @@
 //! Mapping of inputs to sequences of actions.
 use std::collections::HashMap;
 
-use super::actions::Action;
-use super::config;
-
 use serde::Deserialize;
 use strum::IntoEnumIterator;
-use zellij_tile::data::*;
+use zellij_tile::{actions::{Action, Direction}, data::*};
 
 /// Used in the config struct
 #[derive(Clone, Debug, PartialEq)]
