+++ src/common/input/mod.rs
@@ -1,6 +1,5 @@
 //! The way terminal input is handled.
 
-pub mod actions;
 pub mod config;
 pub mod handler;
 pub mod keybinds;
