+++ zellij-tile/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod data;
 pub mod prelude;
 pub mod shim;
+pub mod actions;
 
 use data::*;
